refactor: extract reusable mission domain (#559)
