+++ quality/architecture.toml
@@ -70,6 +70,11 @@ name = "htn-domain-family"
 consumer = "archetype.htn"
 allowed_families = []
 
+[[top_level_family_rule]]
+name = "missions-domain-family"
+consumer = "archetype.missions"
+allowed_families = []
+
 [[package_rule]]
 name = "core-does-not-import-outward"
 consumer = "archetype.core"
@@ -215,7 +220,6 @@ allowed_app = [
   "archetype.app.evaluation.interfaces",
   "archetype.app.missions.interfaces",
   "archetype.app.missions.models",
-  "archetype.app.missions.transitions",
   "archetype.app.query.interfaces",
   "archetype.app.research.interfaces",
   "archetype.app.world.interfaces",
@@ -381,7 +385,6 @@ allowed_interfaces = [
 allowed_app = [
   "archetype.app.artifacts.bundle_models",
   "archetype.app.missions.models",
-  "archetype.app.missions.transitions",
 ]
 
 [[module_rule]]
@@ -577,86 +580,6 @@ owner = "artifacts"
 issue = 558
 reason = "AssetRef is persistent ECS schema still colocated with artifact app authority."
 expiry_condition = "Remove when #558 moves AssetRef to archetype.artifacts.components without duplicating its class identity."
-expires = "v0.5"
-
-[[exception]]
-rule = "app_component_model"
-consumer = "archetype.app.missions.models"
-target = "archetype.app.missions.models.Mission"
-owner = "missions"
-issue = 559
-reason = "Mission is persistent ECS schema still colocated with mission app authority."
-expiry_condition = "Remove when #559 moves Mission to archetype.missions.components without duplicating its class identity."
-expires = "v0.5"
-
-[[exception]]
-rule = "app_component_model"
-consumer = "archetype.app.missions.models"
-target = "archetype.app.missions.models.TaskGate"
-owner = "missions"
-issue = 559
-reason = "TaskGate is persistent ECS schema still colocated with mission app authority."
-expiry_condition = "Remove when #559 moves TaskGate to archetype.missions.components without duplicating its class identity."
-expires = "v0.5"
-
-[[exception]]
-rule = "app_component_model"
-consumer = "archetype.app.missions.models"
-target = "archetype.app.missions.models.Attempt"
-owner = "missions"
-issue = 559
-reason = "Attempt is persistent ECS schema still colocated with mission app authority."
-expiry_condition = "Remove when #559 moves Attempt to archetype.missions.components without duplicating its class identity."
-expires = "v0.5"
-
-[[exception]]
-rule = "app_component_model"
-consumer = "archetype.app.missions.models"
-target = "archetype.app.missions.models.Checkpoint"
-owner = "missions"
-issue = 559
-reason = "Checkpoint is persistent ECS schema still colocated with mission app authority."
-expiry_condition = "Remove when #559 moves Checkpoint to archetype.missions.components without duplicating its class identity."
-expires = "v0.5"
-
-[[exception]]
-rule = "app_component_model"
-consumer = "archetype.app.missions.models"
-target = "archetype.app.missions.models.Finalization"
-owner = "missions"
-issue = 559
-reason = "Finalization is persistent ECS schema still colocated with mission app authority."
-expiry_condition = "Remove when #559 moves Finalization to archetype.missions.components without duplicating its class identity."
-expires = "v0.5"
-
-[[exception]]
-rule = "app_component_model"
-consumer = "archetype.app.missions.models"
-target = "archetype.app.missions.models.Commit"
-owner = "missions"
-issue = 559
-reason = "Commit is persistent ECS schema still colocated with mission app authority."
-expiry_condition = "Remove when #559 moves Commit to archetype.missions.components without duplicating its class identity."
-expires = "v0.5"
-
-[[exception]]
-rule = "app_component_model"
-consumer = "archetype.app.missions.models"
-target = "archetype.app.missions.models.Evidence"
-owner = "missions"
-issue = 559
-reason = "Evidence is persistent ECS schema still colocated with mission app authority."
-expiry_condition = "Remove when #559 moves Evidence to archetype.missions.components without duplicating its class identity."
-expires = "v0.5"
-
-[[exception]]
-rule = "app_component_model"
-consumer = "archetype.app.missions.models"
-target = "archetype.app.missions.models.FrictionLog"
-owner = "missions"
-issue = 559
-reason = "FrictionLog is persistent ECS schema still colocated with mission app authority."
-expiry_condition = "Remove when #559 moves FrictionLog to archetype.missions.components without duplicating its class identity."
 expires = "v0.5"
 
 [[exception]]
