+++ src/archetype/app/missions/__init__.py
@@ -1,7 +1,7 @@
 # Copyright 2026 Vangelis Technologies Inc.
 # SPDX-License-Identifier: Apache-2.0
 
-"""Mission lifecycle and validator-gated state transitions."""
+"""Durable mission claims, authorization, and application orchestration."""
 
 from archetype.app.missions.claim_service import MissionAttemptClaimService
 from archetype.app.missions.execution_service import MissionAttemptExecutionService
@@ -12,55 +12,35 @@ from archetype.app.missions.interfaces import (
     iMissionService,
 )
 from archetype.app.missions.models import (
-    MISSION_COMPONENTS,
-    Attempt,
     AttemptArtifactExpiration,
     AttemptArtifactProjection,
     AttemptArtifactPublication,
     AttemptClaim,
     AttemptClaimAcquisition,
     AttemptRecoveryDecision,
-    Checkpoint,
-    Commit,
-    Evidence,
     FencedAttemptRunner,
     FencedExecutionAuthorization,
-    Finalization,
-    FrictionLog,
-    Mission,
     MissionArtifactFinalizationExpiredError,
     MissionAttemptExecution,
     MissionAttemptRequest,
     PreparedFinalizationSettlement,
     ProviderExecutionCapabilities,
-    TaskGate,
     attempt_invocation_fingerprint,
     mission_attempt_request_fingerprint,
 )
 from archetype.app.missions.service import MissionService
 from archetype.app.missions.transitions import (
     ATTEMPT_CLAIM_TRANSITION_GRAPH,
-    MISSION_TRANSITION_GRAPH,
     AttemptClaimAcquireOutcome,
     AttemptClaimEvent,
     AttemptClaimStatus,
     AttemptClaimTransition,
     AttemptClaimTransitionGraph,
     AttemptRecoveryAction,
-    AttemptStatus,
-    CheckpointStatus,
-    FinalizationPhase,
-    MissionStatus,
-    MissionTaskState,
-    MissionTransition,
-    MissionTransitionEvent,
-    MissionTransitionGraph,
-    TaskStatus,
 )
 
 __all__ = [
-    "MISSION_COMPONENTS",
-    "Attempt",
+    "ATTEMPT_CLAIM_TRANSITION_GRAPH",
     "AttemptArtifactExpiration",
     "AttemptArtifactProjection",
     "AttemptArtifactPublication",
@@ -73,34 +53,16 @@ __all__ = [
     "AttemptClaimTransitionGraph",
     "AttemptRecoveryAction",
     "AttemptRecoveryDecision",
-    "AttemptStatus",
-    "Checkpoint",
-    "CheckpointStatus",
-    "Commit",
-    "Evidence",
-    "Finalization",
-    "FinalizationPhase",
-    "FrictionLog",
     "FencedExecutionAuthorization",
     "FencedAttemptRunner",
-    "Mission",
     "MissionArtifactFinalizationExpiredError",
-    "MissionStatus",
     "MissionAttemptRequest",
     "PreparedFinalizationSettlement",
     "MissionAttemptClaimService",
     "MissionAttemptExecution",
     "MissionAttemptExecutionService",
     "MissionService",
-    "MissionTaskState",
-    "MissionTransition",
-    "MissionTransitionEvent",
-    "MissionTransitionGraph",
-    "MISSION_TRANSITION_GRAPH",
-    "ATTEMPT_CLAIM_TRANSITION_GRAPH",
     "ProviderExecutionCapabilities",
-    "TaskGate",
-    "TaskStatus",
     "iMissionService",
     "iMissionAttemptClaimService",
     "iMissionAttemptExecutionService",
