+++ src/archetype/missions/__init__.py
@@ -1,0 +1,52 @@
+# Copyright 2026 Vangelis Technologies Inc.
+# SPDX-License-Identifier: Apache-2.0
+
+"""Reusable mission ECS components and pure world-state transitions."""
+
+from archetype.missions.components import (
+    MISSION_COMPONENTS,
+    Attempt,
+    Checkpoint,
+    Commit,
+    Evidence,
+    Finalization,
+    FrictionLog,
+    Mission,
+    TaskGate,
+)
+from archetype.missions.transitions import (
+    MISSION_TRANSITION_GRAPH,
+    AttemptStatus,
+    CheckpointStatus,
+    FinalizationPhase,
+    MissionStatus,
+    MissionTaskState,
+    MissionTransition,
+    MissionTransitionEvent,
+    MissionTransitionGraph,
+    TaskStatus,
+    retry_event,
+)
+
+__all__ = [
+    "MISSION_COMPONENTS",
+    "MISSION_TRANSITION_GRAPH",
+    "Attempt",
+    "AttemptStatus",
+    "Checkpoint",
+    "CheckpointStatus",
+    "Commit",
+    "Evidence",
+    "Finalization",
+    "FinalizationPhase",
+    "FrictionLog",
+    "Mission",
+    "MissionStatus",
+    "MissionTaskState",
+    "MissionTransition",
+    "MissionTransitionEvent",
+    "MissionTransitionGraph",
+    "TaskGate",
+    "TaskStatus",
+    "retry_event",
+]
