+++ tests/app/test_mission_transition_authority.py
@@ -11,6 +11,10 @@ from pydantic import ValidationError
 
 from archetype.app.limits import MAX_ICEBERG_SNAPSHOT_ID
 from archetype.app.missions import (
+    MissionService,
+    attempt_invocation_fingerprint,
+)
+from archetype.missions import (
     MISSION_TRANSITION_GRAPH,
     Attempt,
     AttemptStatus,
@@ -18,14 +22,12 @@ from archetype.app.missions import (
     Finalization,
     FinalizationPhase,
     Mission,
-    MissionService,
     MissionStatus,
     MissionTaskState,
     MissionTransitionEvent,
     MissionTransitionGraph,
     TaskGate,
     TaskStatus,
-    attempt_invocation_fingerprint,
 )
